+++ collie_recs/model/deep_fm.py
@@ -11,11 +11,11 @@ from collie_recs.model.base import (BasePipeline,
                                     INTERACTIONS_LIKE_INPUT,
                                     ScaledEmbedding,
                                     ZeroEmbedding)
-from collie_recs.model.base.docstring import merge_docstrings
-from collie_recs.utils import get_init_arguments, trunc_normal
+from collie_recs.utils import get_init_arguments, merge_docstrings, trunc_normal
 
 
 class DeepFM(BasePipeline):
+    # NOTE: the full docstring is merged in with ``BasePipeline``'s using ``merge_docstrings``
     """
     Training pipeline for a deep factorization model.
 
@@ -74,8 +74,6 @@ class DeepFM(BasePipeline):
         on the optimizer
     bias_lr: float
         Bias terms learning rate. If 'infer', will set equal to ``lr``
-    weight_decay: float
-        Weight decay passed to the optimizer, if optimizer permits
     optimizer: torch.optim or str
         If a string, one of the following supported optimizers: