+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
  - filename for ``HybridPretrainedModel`` to ``hybrid_pretrained_matrix_factorization.py``. The former model filepath is now deprecated and will be removed in future version ``0.6.0``
  - ``collie.model.base`` is now split into its own directory with the same name
+ - reduced boilerplate docstrings required for models
  - all ``model.freeze() -> model.eval()``
 
 # [0.4.0] - 2021-5-13